Protect muvm-server with a cookie
While our threat model targets single user systems, let's at least
protect muvm-server with an authentication cookie, so only the user
creating the microVM (and root) can access it.

The cookie is stored in the lockfile, instead of the server port (this
was useless in practice and generated confusion as could override the
command line). The lockfile is located in XDG_RUNTIME_DIR, which in
properly configured systems is only accessible to its owner.

Signed-off-by: Sergio Lopez <slp@redhat.com>